Move most user apps into Home Manager
Those left behind are either more integrated into the system and will be left for a potential NixOS migration, or have some sort of bug such as that linked in the commit with Rofi. A number of missing apps are no longer system-wide and instead live in project-specific shell.nix configs.