Switch to 23.05 stable (pre-release)
I can switch back to unstable if I need something like mesa more up to
date. The issue at the moment is that, with the derivation cache being
busted so easily, every update to nixpkgs causes a multi-gigabyte
download, which on my trash internet is annoying. Case in point, this
input update clocks in at a 3.5GB download. I'm _hoping_ that switching
to a stable release branch will reduce this a bit.