Add light/dark system theme support
`extraPortals` is managing dbus, systemd, etc for us. If darkman isn't
installed like this then it'll work within itself but fail to set the
below global dbus value.

This works immediately for Firefox via `darkman toggle`. It sets the
following dbus setting:

$ nix-shell -p qt6.qttools.dev
$ $ qdbus org.freedesktop.portal.Desktop /org/freedesktop/portal/desktop org.freedesktop.portal.Settings.Read "org.freedesktop.appearance" "color-scheme"

Light and dark set the value to 1 or 2. 0 denotes no preference and is
the default. Without a portal providing the setting(s) there'd be an error in
the above command that's also visible with:

$ MOZ_LOG=LookAndFeel:5 firefox