Move dev terminal back to macOS
And move from Neovim to Helix, and drop deploy-rs which has been
broken for a little bit.

I don't have the energy for all the configuration at the moment. I'm
still using NixOS for the homelab (Tentacool), and still love
configuring my dev/shell environment in Nix; it made this move so, so
much easier.

The flake is currently written pretty lazily, assuming I'll only need
the shell or packages in this OS/arch. I can improve that later.