Set static ULA
The server had to be restarted for firewall changes (such as allowing pings) to
take effect. After that the Blocky service needed to be restarted as it was
initially unable to resolve upstream.

Where in IPv4 the DHCP server (router) assigns a fixed IP, in IPv6 the device
self-assigns IPs.

ULAs are as far as I'm aware appropriate for local networking.

This particular ULA prefix appears to have been generated by UniFi and is
consistent across devices, however it's unclear if it's stable. For now UniFi's
IPv6 DNS is pointed here, equivalent to IPv4. As expected the ULA prefix is
different from my ISP-supplied global prefix.

It can be tested from clients with:
$ dig -6 doubleclick.net

Or more directly with:
$ dig doubleclick.net @fd1d:8607:ebdd:b24b::1