Add SSH key to agent on shell init
Fixes jj/git usage of the key before explicitly using ssh-add or connecting to a
server or something with it.

This costs ~15ms. Backgrounding it prevents perceptible harm to shell startup.
-l may be marginally faster following the initial add but it's not worth the
complexity.

stderr is redirected to avoid shell startup noise.

The IdentityFile continues to be used in the absence of a key path.

This only works because my key has no passphrase. My threat model doesn't
account for the risk of someone gaining physical access to my machine,
particularly with macOS security in front of it.