Switch to unstable
Stable is compelling for configuration-heavy stuff that I don't want to touch
often like NixOS servers, however that no longer applies. It's now easier to
track unstable across the board given e.g. I want the latest version of Jujutsu
(and it needs to correlate to Home Manager to avoid annoying mismatch warnings).

As of this commit we're slightly ahead of Hydra for the sake of jj 0.29.