Add initial jjui config
There's overlap between the jj and jjui configs, but they're not 1:1 as e.g. we
don't need a tug alias equivalent in jjui as that's the default option under the
bookmarks menu.

I'm a bit unclear on leader vs custom commands, and on how $revset would work
instead of $change_id.