Home Manager supports null packages
This solves the compat issue between Ghostty and Bat since the module checks if
the package is nullable. Previously it'd be non-nullable and then the symlink
would point to somewhere that doesn't exist in the Nix store. We'll still be
missing syntax highlighting for Ghostty's config, but as per use of Home Manager
that won't matter much!